pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fixos"
version = "2.1.0"
description = "AI-powered Linux/Windows diagnostics and repair – audio, hardware, system issues"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "fixos contributors" },
    { name = "Tom Sapletta", email = "tom@sapletta.com" },
]
keywords = ["linux", "windows", "diagnostics", "ai", "llm", "audio", "system-repair", "cross-platform"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Monitoring",
    "Environment :: Console",
]
dependencies = [
    "openai>=1.35.0",
    "prompt_toolkit>=3.0.43",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
fixos = "fixos.cli:main"

[project.urls]
Homepage = "https://github.com/wronai/fixos"
"Bug Tracker" = "https://github.com/wronai/fixos/issues"

[tool.setuptools.packages.find]
exclude = ["tests*", "docker*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["fixos"]
omit = ["tests/*", "docker/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
